refactor(table): rename to table-widget
