+++ packages/components/_components-themes.scss
@@ -10,7 +10,7 @@
 @import 'multilanguage/xm-angular-editor-theme';
 @import 'navbar/search/xm-navbar-search-widget-theme';
 @import 'sidebar/sidebar-theme';
-@import 'table/table/xm-table-theme';
+@import 'table/table-widget/xm-table-widget-theme';
 
 @mixin xm-components-theme($xm-theme) {
   @include xm-sidebar-theme($xm-theme);
