+++ packages/components/table/index.ts
@@ -18,5 +18,5 @@ export {
     XmTableDynamicColumnsCellComponent,
 } from './columns/xm-table-dynamic-columns-cell.component';
 
-export { XmTableComponent,
-} from './table/xm-table.component';
+export { XmTableWidget,
+} from '@xm-ngx/components/table/table-widget/xm-table-widget.component';
