+++ packages/components/table/table-widget/ng-package.json
