+++ packages/components/table/table-widget/xm-table-widget.component.scss
